requirements.txt
datashader==0.12.1
holoviews==1.14.2
matplotlib==3.3.1
pandas=1.2.3
progress==1.5
PyYAML==5.4.1
scikit-image==0.18.1
torch==1.7.1
torchvision==0.8.2
umap-learn==0.5.1
